+++ portality/events/consumers/application_publisher_created_notify.py
@@ -1,5 +1,5 @@
 # from flask import url_for
-from portality.lib.flask import url_for
+from portality.util import url_for
 
 from portality.lib import dates
 from portality.events.consumer import EventConsumer
