+++ portality/events/consumers/application_publisher_inprogress_notify.py
@@ -1,5 +1,5 @@
 # from flask import url_for
-from portality.lib.flask import url_for
+from portality.util import url_for
 
 from portality.core import app
 from portality.events.consumer import EventConsumer
