+++ portality/events/consumers/journal_assed_assigned_notify.py
@@ -1,5 +1,5 @@
 # from flask import url_for
-from portality.lib.flask import url_for
+from portality.util import url_for
 
 from portality.events.consumer import EventConsumer
 from portality import constants
